Fixed a crash when a surface with a string object was unloaded. Fixed a rendering issue when no TextFormat option was specified. Fixed a shadow rendering issue when certain fonts were used.

+++ Semper/include/bind.h
@@ -44,7 +44,7 @@ void bind_set_raw(void* soi, unsigned char* pn, unsigned char* pv);
 void bind_reset(object* o);
 
 void bind_unbind(surface_data* sd, unsigned char* sn);
-int bind_update_string(object* o, string_bind* sb);
+size_t  bind_update_string(object* o, string_bind* sb);
 int bind_update_numeric(object* o, bind_numeric* bn);
 double bind_percentual_value(double val, double min_val, double max_val);
 unsigned char *bind_source_name(object *o,size_t index);
